+++ packages/expo-passkeys/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @clerk/expo-passkeys
 
+## 0.3.36
+
+### Patch Changes
+
+- Updated dependencies [[`600c648`](https://github.com/clerk/javascript/commit/600c648d4087a823341041c90018797fbc0033f0)]:
+  - @clerk/shared@3.23.0
+  - @clerk/types@4.83.0
+
 ## 0.3.35
 
 ### Patch Changes