+++ packages/react-router/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.9.6
+
+### Patch Changes
+
+- Updated dependencies [[`600c648`](https://github.com/clerk/javascript/commit/600c648d4087a823341041c90018797fbc0033f0)]:
+  - @clerk/shared@3.23.0
+  - @clerk/clerk-react@5.45.0
+  - @clerk/types@4.83.0
+  - @clerk/backend@2.10.1
+
 ## 1.9.5
 
 ### Patch Changes